Module drm::mode

Structs

Connector
A connector represents the final output.

Crtc
A CRTC the first stage in the encoding pipeline.

DumbBuf

DumbBufFlags

Encoder

Fb
Fb is the DRM representation of frame buffer.

Id
Id is a 32-bit integer that represents an object in the driver.

ModeFlags

ModeInfo
The mode configuration fo the screen.

ModeType

Plane
A plane is an extra frame buffer that the CRTC's can compose onto the output during scanout.

Property
A property of a connector.

PropertyBlob
Blob's are used to get EDID information out of a Property.

PropertyEnum

PropertyFlags

Props

Resources
Resource lists all the Resources avaible on the card.

Enums

Connection

ConnectorType
Types of connectors.

EncoderType

Subpixel

Constants

DUNNO

FLAG_3D_FIELD_ALTERNATIVE

FLAG_3D_FRAME_PACKING

FLAG_3D_LINE_ALTERNATIVE

FLAG_3D_L_DEPTH

FLAG_3D_L_DEPTH_GFX_GFX_DEPTH

FLAG_3D_NONE

FLAG_3D_SIDE_BY_SIDE_FULL

FLAG_3D_SIDE_BY_SIDE_HALF

FLAG_3D_TOP_AND_BOTTOM

FLAG_BCAST

FLAG_CLKDIV2

FLAG_CSYNC

FLAG_DBLCLK

FLAG_DBLSCAN

FLAG_HSKEW

FLAG_INTERLACE

FLAG_NCSYNC

FLAG_NHSYNC

FLAG_NVSYNC

FLAG_PCSYNC

FLAG_PHSYNC

FLAG_PIXMUX

FLAG_PVSYNC

PROP_BITMASK

PROP_BLOB

PROP_ENUM

PROP_IMMUTABLE

PROP_PENDING

PROP_RANGE

TYPE_BUILTIN

TYPE_CLOCK_C

TYPE_CRTC_C

TYPE_DEFAULT

TYPE_DRIVER

TYPE_PREFERRED

TYPE_USERDEF

Traits

Resource