Struct drm::mode::Connector

pub struct Connector { /* fields omitted */ }

A connector represents the final output.

Usually a physical display connected by a cable.

Methods

impl Connector

fn props<'a>(&'a self) -> Props<'a>

Get an iterator over the properties of this connector and the values they have been set to.

fn modes(&self) -> &[ModeInfo]

List of avaible modes on this connector.

fn possible_encoders(&self) -> &[Id<Encoder>]

List of encoders that can be used with this connector.

fn encoder_id(&self) -> Option<Id<Encoder>>

Show the current encoder being used.

fn connector_id(&self) -> Id<Connector>

Get the id of this encoder.

fn connector_type(&self) -> ConnectorType

Get the type of connector.

fn connection(&self) -> Connection

Get the state of the connector. (Connected/Disconnected).

fn subpixel(&self) -> Subpixel

Get the subpixel layout. (The positions the colors are layed out on the screen).