Struct drm::mode::Plane

pub struct Plane { /* fields omitted */ }

A plane is an extra frame buffer that the CRTC's can compose onto the output during scanout.

Old video game consoles performed rendering by using Sprites and composing them onto the scene using specialised hardware.

Vidoe devices sitll have this hardware. A common use is for hardware cursors. (DRM has apis for Cursors, it's in the TODO list).

Intel hardware tends to support 3 non-transparent planes. (So can't be used for cursors). Assuming the planes can be used to compose video content onto the screen. (Possibly PAVP?)

Methods

impl Plane

fn get_ids(dev: &Device) -> Result<Vec<u32>>

Planes are not listed in the Resources structure so we must fetch a list of them seperately.

fn get(dev: &Device, id: u32) -> Result<Plane>