Struct drm::Master

pub struct Master<'a> { /* fields omitted */ }

Represents a card when the process is the card master.

Only one process can control modesetting of a card. To do this they use a DRM_IOCTL_SET_MASTER. If it successful modesetting, calls can be performed until DRM_IOCTL_UNSET_MASTER, is called or the process exits.

This object is created when Device::set_master is called and when dropped, it will make the Unset call.

Due to some mutability constraits this also implements DerefMut for the associated card. (Essentailly it used to be impossible to read events and have card master at the same time, because Read requires mutable access).

Methods

impl<'a> Master<'a>

fn set_crtc(&self,
crtc_id: Id<Crtc>,
fb: Option<Id<Fb>>,
x: u32,
y: u32,
connectors: &[Id<Connector>],
mode: Option<&ModeInfo>)
-> Result<()>

Set the mode, and the frame buffer.

The driver will pick a path from the given crtc to the given connectors.