Crate drm

Modules

mode

Structs

Device
Open fd for a dri device. Such as /dev/dri/card0.

Master
Represents a card when the process is the card master.

Version
Version information.

Enums

Event
Event types read from device. Event's are sent only after a request vblank has been called or a page flip has been issued.

EventType
Type of a drm device event.