Module drm::control::property

Property

A property of a modesetting resource.

All modesetting resources have a set of properties that have values that can be modified. These properties are modesetting resources themselves, and may even have their own set of properties.

Properties may have mutable values attached to them. These can be changed by either changing the state of a resource (thereby affecting the property), directly changing the property value itself, or by batching property changes together and executing them all atomically.

Structs

EnumValue

A single value of ValueType::Enum type

EnumValues

A set of EnumValues for a single property

Handle

A handle to a property

Info

Information about a property

Enums

Value

The value of a property, in a typed format

ValueType

Describes the types of value that a property uses.

Type Definitions

RawValue

A raw property value that does not have a specific property type