drm

Module node

Module for abstractions on drm device nodes.

Modules

constants

OS-Specific DRM constants.

Structs

DrmNode

A node which refers to a DRM device.

Enums

CreateDrmNodeError

An error that may occur when creating a DrmNode from a file descriptor.

NodeType

A type of node

Functions

dev_path

Returns the path of a specific type of node from the DRM device described by major and minor device numbers.

is_device_drm

Returns if the given device by major:minor pair is a DRM device.

node_path

Returns the path of a specific type of node from the same DRM device as an existing DrmNode.

path_to_type

Returns the path of a specific type of node from the same DRM device as another path of the same node.